requirements.txt
# requirements.txt
# This file contains the exact package versions for a stable environment.
# Install all packages using: pip install -r requirements.txt

# Core ML Framework
torch
torchvision

# Data Handling & Utilities
numpy<2.0
pandas
Pillow
tqdm

# Model Conversion (PyTorch -> TFLite)
onnx==1.12.0
onnx-tf==1.10.0
tensorflow==2.10.1
tensorflow-probability==0.18.0
protobuf==3.20.3

# Web Application
Flask